% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisser.R
\name{bs_theme_esquisse}
\alias{bs_theme_esquisse}
\title{Bootstrap Theme for Esquisse}
\usage{
bs_theme_esquisse()
}
\value{
A \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.
}
\description{
Bootstrap Theme for Esquisse
}
