% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{esquisse-deprecated}
\alias{esquisse-deprecated}
\alias{esquisseContainer}
\title{Deprecated functions}
\usage{
esquisseContainer(...)
}
\arguments{
\item{...}{See \code{\link[=esquisse_container]{esquisse_container()}}}
}
\description{
Deprecated functions
}
\note{
The following functions are deprecated and will be removed in next release:
\itemize{
\item \code{esquisseContainer} : replaced by \code{esquisse_container}
}
}
