% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisserServer.R, R/esquisserUI.R
\name{esquisserServer}
\alias{esquisserServer}
\alias{esquisse-module}
\alias{esquisserUI}
\title{Esquisse Shiny module}
\usage{
esquisserServer(input, output, session, data = NULL)

esquisserUI(id, header = TRUE, choose_data = TRUE)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{a \code{reactiveValues} with at least a slot \code{data} containing a \code{data.frame}
to use in the module. And a slot \code{name} corresponding to the name of the \code{data.frame}.}

\item{id}{Module's id.}

\item{header}{Logical. Display or not \code{esquisse} header.}

\item{choose_data}{Logical. Display or not the button to choose data.}
}
\description{
Launch \code{esquisse} in a classic Shiny app.
}
