% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse-package.R
\docType{package}
\name{esquisse}
\alias{esquisse}
\alias{esquisse-package}
\title{Explore and Visualize Your Data Interactively}
\description{
A 'shiny' gadget to create 'ggplot2' figures interactively with drag-and-drop to map your variables to different aesthetics.
You can quickly visualize your data accordingly to their type, export in various formats,
and retrieve the code to reproduce the plot.
}
\examples{
\dontrun{

esquisser()

# launch esquisse with specific data:
esquisser(mtcars)

}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dreamrs.github.io/esquisse/}
  \item \url{https://github.com/dreamRs/esquisse}
  \item Report bugs at \url{https://github.com/dreamRs/esquisse/issues}
}

}
\author{
Fanny Meyer & Victor Perrier (@dreamRs_fr)
}
