% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisser.R
\name{esquisser}
\alias{esquisser}
\title{An addin to easily create chart with ggplot2}
\usage{
esquisser(data = NULL)
}
\arguments{
\item{data}{a data.frame, you can pass a data.frame explicitly to the function, 
otherwise you'll have to choose one in your environment.}
}
\value{
code to reproduce chart
}
\description{
An addin to easily create chart with ggplot2
}
\examples{
\dontrun{
esquisser(iris)
}


}
