% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivoter.R
\name{pivoter}
\alias{pivoter}
\title{Addin to aid with visualizing the result of pivoting with `tidyr`.}
\usage{
pivoter(data = NULL, coerceVars = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} passed by the user, or one selected from
the global environment.}

\item{coerceVars}{Logical. Variables can be coerced to a different type
when selecting data. Defaults to \code{FALSE}.}
}
\value{
Code to reproduce the pivoted dataframe.
}
\description{
Addin to aid with visualizing the result of pivoting with `tidyr`.
}
