% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotServer.R
\name{pivotServer}
\alias{pivotServer}
\title{Server associated with pivot Shiny module.}
\usage{
pivotServer(input, output, session, data = NULL,
  dataModule = c("GlobalEnv", "ImportFile"), sizeDataModule = "m")
}
\arguments{
\item{input, output, session}{Standards \code{shiny} server arguments.}

\item{data}{A \code{reactiveValues} with at least a slot \code{data} 
containing a \code{data.frame} to use in the module. And a slot 
\code{name} corresponding to the name of the \code{data.frame}.}

\item{dataModule}{Data module to use, choose between \code{"GlobalEnv"}
or \code{"ImportFile"}.}

\item{sizeDataModule}{Size for the modal window for selecting data.}
}
\description{
Server associated with pivot Shiny module.
}
