% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotUI.R
\name{pivotUI}
\alias{pivotUI}
\title{Pivot Shiny module.}
\usage{
pivotUI(id, container = esquisseContainer(), insert_code = TRUE)
}
\arguments{
\item{id}{Module id.}

\item{container}{Defaults to using the \code{esquisse} container as the
addin display.}

\item{insert_code}{Logical. Displays a button to insert \pkg{tidyr} code
in the current user script. Defaults to \code{TRUE}.}
}
\value{
Reactive values with 2 slots containing the code to generate
  the pivot table result and the resulting dataframe.
}
\description{
Pivot Shiny module.
}
