% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_module.R
\name{run_module}
\alias{run_module}
\title{Run module example}
\usage{
run_module(module = c("filterDF", "chooseData", "chooseData2", "coerce"))
}
\arguments{
\item{module}{Module for which to see a demo.}
}
\description{
Run module example
}
\examples{

if (interactive()) {

# Demo for filterDF module
run_module("filterDF")

}

}
