% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_call.R
\name{tidyr_call}
\alias{tidyr_call}
\title{Make a tidyr call}
\usage{
tidyr_call(data = NULL, targets = NULL, settings = NULL)
}
\arguments{
\item{data}{A \code{reactiveValues} with at least a slot \code{data} 
containing a \code{data.frame} to use in the module. And a slot 
\code{name} corresponding to the name of the \code{data.frame}.}

\item{targets}{Desired variables to pivot on in the \code{data.frame}.}

\item{settings}{Desired pivot settings as selected by the user.}
}
\value{
An expression that can be parsed.
}
\description{
Make a tidyr call
}
