% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse-package.R
\docType{package}
\name{esquisse}
\alias{esquisse}
\title{Explore and Visualize Your Data Interactively}
\description{
A 'shiny' gadget to create 'ggplot2' figures interactively with drag-and-drop to map your variables to different aesthetics.
You can quickly visualize your data accordingly to their type, export in various formats,
and retrieve the code to reproduce the plot.
}
\examples{
\dontrun{

esquisser()

# launch esquisse with specific data:
esquisser(mtcars)

}
}
\author{
Fanny Meyer & Victor Perrier (@dreamRs_fr)
}
