% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisser.R
\name{esquisser}
\alias{esquisser}
\title{An add-in to easily create plots with ggplot2}
\usage{
esquisser(
  data = NULL,
  coerce_vars = getOption(x = "esquisse.coerceVars", default = TRUE),
  disable_filters = getOption(x = "esquisse.disable_filters", default = FALSE),
  viewer = getOption(x = "esquisse.viewer", default = "dialog")
)
}
\arguments{
\item{data}{a data.frame, you can pass a data.frame explicitly to the function,
otherwise you'll have to choose one in global environment.}

\item{coerce_vars}{If \code{TRUE} allow to coerce variables to different type when selecting data.}

\item{disable_filters}{Logical. Disable the menu allowing to filter data used.}

\item{viewer}{Where to display the gadget: \code{"dialog"},
\code{"pane"} or \code{"browser"} (see \code{\link[shiny]{viewer}}).}
}
\value{
code to reproduce chart.
}
\description{
An add-in to easily create plots with ggplot2
}
\examples{
if (interactive()) {
# Launch with :
esquisser(iris)
# If in RStudio it will be launched by default in dialog window
# If not, it will be launched in browser

# Launch esquisse in browser :
esquisser(iris, viewer = "browser")

# You can set this option in .Rprofile :
options("esquisse.viewer" = "viewer")
# or
options("esquisse.viewer" = "browser")

# esquisse use shiny::runApp
# see ?shiny::runApp to see options
# available, example to use custom port:

options("shiny.port" = 8080)
esquisser(iris, viewer = "browser")

}
}
