% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{module-chooseData}
\alias{module-chooseData}
\alias{chooseDataUI}
\alias{chooseDataServer}
\title{Module for choosing data.frame}
\usage{
chooseDataUI(id, label = "Data", icon = "database", width = "100\%", ...)

chooseDataServer(
  input,
  output,
  session,
  dataModule = c("GlobalEnv", "ImportFile"),
  data = NULL,
  name = NULL,
  selectVars = TRUE,
  selectedTypes = c("continuous", "discrete", "time"),
  coerceVars = FALSE,
  launchOnStart = TRUE,
  size = "m"
)
}
\arguments{
\item{id}{Module's id.}

\item{label}{Label for button, passed to \code{\link[shiny:actionButton]{actionButton}}.}

\item{icon}{Icon to appears on the button, passed to \code{\link[shiny:actionButton]{actionButton}}.}

\item{width}{Width of button, passed to \code{\link[shiny:actionButton]{actionButton}}.}

\item{...}{Other arguments passed to \code{\link[shiny:actionButton]{actionButton}}}

\item{input, output, session}{standards \code{shiny} server arguments.}

\item{dataModule}{Data module to use, choose between \code{"GlobalEnv"}
(select ad \code{data.frame} from Global environment)
or \code{"ImportFile"} (import an external file supported by \code{\link[rio]{import}}).}

\item{data}{A \code{data.frame} to use by default.}

\item{name}{Character, object's name to use for \code{data}.}

\item{selectVars}{Display module to select variables, \code{TRUE} by default.}

\item{selectedTypes}{Type of variables selected by default in select variables module.
Possible types are \code{"discrete"}, \code{"time"}, \code{"continuous"} and \code{"id"},
by default \code{"id"} is discarded.}

\item{coerceVars}{Display module to coerce variables between different class, \code{TRUE} by default.}

\item{launchOnStart}{Opens modal window when the application starts.}

\item{size}{Size for the modal window.}
}
\value{
a \code{\link[shiny]{reactiveValues}} containing the data selected under slot \code{data}
and the name of the selected \code{data.frame} under slot \code{name}.
}
\description{
DEPRECATED, please see package \href{https://github.com/dreamRs/datamods}{datamods} for similar features.
}
