% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{module-coerce}
\alias{module-coerce}
\alias{coerceUI}
\alias{coerceServer}
\title{Coerce data.frame's columns module}
\usage{
coerceUI(id)

coerceServer(input, output, session, data, reactiveValuesSlot = "data")
}
\arguments{
\item{id}{Module id. See \code{\link[shiny]{callModule}}.}

\item{input, output, session}{standards \code{shiny} server arguments.²}

\item{data}{A \code{data.frame} or a \code{reactive}
function returning a \code{data.frame} or a
\code{reactivevalues} with a slot containing a \code{data.frame}
(use \code{reactiveValuesSlot} to identify that slot)}

\item{reactiveValuesSlot}{If \code{data} is a \code{reactivevalues},
specify the name of the slot containing data.}
}
\value{
a \code{reactiveValues} with two slots: \code{data} original \code{data.frame}
with modified columns, and \code{names} column's names with call to coerce method.
}
\description{
DEPRECATED, please see package \href{https://github.com/dreamRs/datamods}{datamods} for similar features.
}
